refactor(privval,p2p): share low-order point check, assert handshake test errors
Move the duplicated Curve25519 low-order point list into a single internal
package (internal/libs/x25519) consumed by both the privval and p2p
secret-connection handshakes, preserving the existing early reject of such
ephemeral keys. Capture and assert the handshake script error in the affected
tests so a failing script no longer passes silently.


<sub>🤖 Co-authored by [Claudius the Magnificent](https://github.com/lklimek/claudius) AI Agent</sub>
