+++ privval/secret_connection.go
@@ -5,7 +5,7 @@ import (
 	"crypto/cipher"
 	crand "crypto/rand"
 	"crypto/sha256"
-	"crypto/subtle"
+
 	"encoding/binary"
 	"errors"
 	"fmt"
@@ -29,6 +29,7 @@ import (
 	"github.com/dashpay/tenderdash/crypto/encoding"
 	"github.com/dashpay/tenderdash/internal/libs/async"
 	"github.com/dashpay/tenderdash/internal/libs/protoio"
+	"github.com/dashpay/tenderdash/internal/libs/x25519"
 	tmprivval "github.com/dashpay/tenderdash/proto/tendermint/privval"
 )
 
@@ -116,7 +117,7 @@ func MakeSecretConnection(conn io.ReadWriteCloser, locPrivKey crypto.PrivKey) (*
 
 	// Reject remote ephemeral keys that decode to a known low-order point
 	// before they are used in the key exchange.
-	if isLowOrderPoint(remEphPub) {
+	if x25519.IsLowOrderPoint(remEphPub) {
 		return nil, ErrSmallOrderRemotePubKey
 	}
 
@@ -383,49 +384,6 @@ func deriveSecrets(
 	}
 
 	return
-}
-
-// lowOrderPoints lists the Curve25519 points whose order is too small to be
-// safe in a Diffie-Hellman exchange. The encodings match the canonical
-// libsodium blocklist; the high bit of the final byte is ignored during
-// comparison, so both encodings of each point are covered.
-var lowOrderPoints = [][32]byte{
-	// 0 (order 4)
-	{0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00},
-	// 1 (order 1)
-	{0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00},
-	// order 8
-	{0xe0, 0xeb, 0x7a, 0x7c, 0x3b, 0x41, 0xb8, 0xae, 0x16, 0x56, 0xe3, 0xfa, 0xf1, 0x9f, 0xc4, 0x6a,
-		0xda, 0x09, 0x8d, 0xeb, 0x9c, 0x32, 0xb1, 0xfd, 0x86, 0x62, 0x05, 0x16, 0x5f, 0x49, 0xb8, 0x00},
-	// order 8
-	{0x5f, 0x9c, 0x95, 0xbc, 0xa3, 0x50, 0x8c, 0x24, 0xb1, 0xd0, 0xb1, 0x55, 0x9c, 0x83, 0xef, 0x5b,
-		0x04, 0x44, 0x5c, 0xc4, 0x58, 0x1c, 0x8e, 0x86, 0xd8, 0x22, 0x4e, 0xdd, 0xd0, 0x9f, 0x11, 0x57},
-	// p-1 (order 2)
-	{0xec, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-		0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x7f},
-	// p (order 4)
-	{0xed, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-		0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x7f},
-	// p+1 (order 1)
-	{0xee, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-		0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x7f},
-}
-
-// isLowOrderPoint reports whether p decodes to one of the known low-order
-// Curve25519 points. The comparison is constant-time and masks the high bit of
-// the final byte, mirroring the canonical libsodium check.
-func isLowOrderPoint(p *[32]byte) bool {
-	var candidate [32]byte
-	copy(candidate[:], p[:])
-	candidate[31] &= 0x7f
-
-	var found int
-	for i := range lowOrderPoints {
-		found |= subtle.ConstantTimeCompare(candidate[:], lowOrderPoints[i][:])
-	}
-	return found == 1
 }
 
 // computeDHSecret computes a Diffie-Hellman shared secret key
